requirements.txt
# Enhanced PAD Document Semantic Matcher - Requirements
streamlit>=1.49.0
pandas>=2.0.0
numpy>=1.24.0
nltk>=3.8.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-groq>=0.3.0
langchain-mistralai>=0.3.0
mistralai>=1.0.0
openai>=1.0.0
watchdog>=3.0.0